+++ internal/handlers/pg/handler.go
@@ -19,20 +19,17 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"sort"
 	"sync/atomic"
 	"time"
 
 	"github.com/AlekSi/pointer"
 	"github.com/prometheus/client_golang/prometheus"
 	"go.uber.org/zap"
-	"golang.org/x/exp/maps"
 
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
 	"github.com/FerretDB/FerretDB/internal/handlers/pg/pgdb"
 	"github.com/FerretDB/FerretDB/internal/types"
 	"github.com/FerretDB/FerretDB/internal/util/lazyerrors"
-	"github.com/FerretDB/FerretDB/internal/util/must"
 	"github.com/FerretDB/FerretDB/internal/wire"
 )
 
@@ -115,7 +112,7 @@ func (h *Handler) Handle(ctx context.Context, reqHeader *wire.MsgHeader, reqBody
 		requests.WithLabelValues(command).Inc()
 
 		resHeader.OpCode = wire.OP_REPLY
-		resBody, err = h.handleOpQuery(ctx, query, command)
+		resBody, err = h.MsgQueryCmd(ctx, query)
 
 	case wire.OP_REPLY:
 		fallthrough
@@ -213,23 +210,6 @@ func (h *Handler) handleOpMsg(ctx context.Context, msg *wire.OpMsg, cmd string) 
 	return nil, common.NewErrorMsg(common.ErrCommandNotFound, errMsg)
 }
 
-func (h *Handler) handleOpQuery(ctx context.Context, query *wire.OpQuery, cmd string) (*wire.OpReply, error) {
-	if query.FullCollectionName == "admin.$cmd" {
-		return h.QueryCmd(ctx, query)
-	}
-
-	msg := fmt.Sprintf("handleOpQuery: unhandled collection %q", query.FullCollectionName)
-	return nil, common.NewErrorMsg(common.ErrNotImplemented, msg)
-}
-
-// MsgQueryCmd used for debugging purposes.
-// XXX: now it cannot be called from handleOpMsg
-// but the presence of the function is necessary to implement common.Handler interface
-// in the process of extracting common handler interface.
-func (h *Handler) MsgQueryCmd(ctx context.Context, query *wire.OpQuery) (*wire.OpReply, error) {
-	panic("unreachable code")
-}
-
 // MsgDebugError used for debugging purposes.
 func (h *Handler) MsgDebugError(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	return nil, errors.New("debug_error")
@@ -238,30 +218,4 @@ func (h *Handler) MsgDebugError(ctx context.Context, msg *wire.OpMsg) (*wire.OpM
 // MsgDebugPanic used for debugging purposes.
 func (h *Handler) MsgDebugPanic(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
 	panic("debug_panic")
-}
-
-// MsgListCommands returns a list of currently supported commands.
-func (h *Handler) MsgListCommands(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, error) {
-	var reply wire.OpMsg
-
-	cmdList := must.NotFail(types.NewDocument())
-	names := maps.Keys(common.Commands)
-	sort.Strings(names)
-	for _, name := range names {
-		cmdList.Set(name, must.NotFail(types.NewDocument(
-			"help", common.Commands[name].Help,
-		)))
-	}
-
-	err := reply.SetSections(wire.OpMsgSection{
-		Documents: []*types.Document{must.NotFail(types.NewDocument(
-			"commands", cmdList,
-			"ok", float64(1),
-		))},
-	})
-	if err != nil {
-		return nil, lazyerrors.Error(err)
-	}
-
-	return &reply, nil
 }
